+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 #### Upcoming Changes
 
-* feat(breaking): Replace `cairo-felt` crate with `stark-felt` [#1408](https://github.com/lambdaclass/cairo-vm/pull/1408)
+* feat(BREAKING): Replace `cairo-felt` crate with `starknet-types-core` `Felt` [#1408](https://github.com/lambdaclass/cairo-vm/pull/1408)
 
-* feat/BREAKING: Add Cairo 1 proof mode compilation and execution [#1517] (https://github.com/lambdaclass/cairo-vm/pull/1517)
+* feat(BREAKING): Add Cairo 1 proof mode compilation and execution [#1517] (https://github.com/lambdaclass/cairo-vm/pull/1517)
     * In the cairo1-run crate, now the Cairo 1 Programs are compiled and executed in proof-mode
     * BREAKING: Remove `CairoRunner.proof_mode: bool` field and replace it with `CairoRunner.runner_mode: RunnerMode`