+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 * refactor: remove static lifetime for name str parameter requirement for constant getter
 
-* feat(BREAKING): Replace `cairo-felt` crate with `starknet-types-core` [#1408](https://github.com/lambdaclass/cairo-vm/pull/1408)
+* feat(BREAKING): Replace `cairo-felt` crate with `starknet-types-core` (0.0.5) [#1408](https://github.com/lambdaclass/cairo-vm/pull/1408)
 
 * feat(BREAKING): Add Cairo 1 proof mode compilation and execution [#1517] (https://github.com/lambdaclass/cairo-vm/pull/1517)
     * In the cairo1-run crate, now the Cairo 1 Programs are compiled and executed in proof-mode